+++ lib/src/model/entity.dart
@@ -6,6 +6,7 @@ class Entity {
   final bool generateEntity;
   final List<String> dependencies = [];
   final List<Field> fields = [];
+  final List<String> serializers = [];
 
   Entity({required this.name, required this.copyWith, required this.generateEntity});
 }