+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 0.0.2
 
-- Feature: skip entity class creation on demand, only serialization generation (attribute generateEntity="false")
+- Feature: select which serializers are used per entity (attribute serializers on class node)
+- Feature: skip entity class creation on demand, only serialization generation (attribute generateEntity="false" on class node)
 - Feature: custom imports defined on xml level (import node)
 - Feature: type now can be given in xml node name, no need to use field node
-- Feature: add comment to generated field (attribute comment)
+- Feature: add comment to generated field (attribute comment on field node)
 - Improvement: More helper functions (duration)
 
 ## 0.0.1